refactor(Constants): 상수 추가
